--- BukkitStaffChat/src/main/java/dev/imlukas/bukkitstaffchat/database/impl/sql/functions/ResultSetConsumer.java
@@ -1,9 +1,0 @@
-package dev.imlukas.bukkitstaffchat.database.impl.sql.functions;
-
-import java.sql.ResultSet;
-
-@FunctionalInterface
-public interface ResultSetConsumer {
-
-    void accept(ResultSet resultSet) throws Exception;
-}
