--- BukkitStaffChat/src/main/java/dev/imlukas/bukkitstaffchat/database/impl/sql/functions/ResultSetFunction.java
@@ -1,10 +1,0 @@
-package dev.imlukas.bukkitstaffchat.database.impl.sql.functions;
-
-import java.sql.ResultSet;
-
-@FunctionalInterface
-public interface ResultSetFunction<T> {
-
-    T apply(ResultSet resultSet) throws Exception;
-
-}
