--- BukkitStaffChat/src/main/java/dev/imlukas/bukkitstaffchat/database/impl/sql/impl/MySQLDatabase.java
@@ -1,33 +1,0 @@
-package dev.imlukas.bukkitstaffchat.database.impl.sql.impl;
-
-import dev.imlukas.bukkitstaffchat.database.impl.sql.AbstractRemoteSqlDatabase;
-
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.util.Properties;
-
-public abstract class MySQLDatabase extends AbstractRemoteSqlDatabase {
-    @Override
-    public String getName() {
-        return "mysql";
-    }
-
-    @Override
-    protected Connection createConnection() {
-        try {
-            Class.forName("com.mysql.jdbc.Driver");
-
-            Properties properties = new Properties();
-
-            properties.setProperty("user", username);
-            properties.setProperty("password", password);
-            properties.setProperty("useSSL", "false");
-            properties.setProperty("autoReconnect", "true");
-
-            return DriverManager.getConnection("jdbc:mysql://" + host + ":" + port + "/" + database, properties);
-        } catch (Exception ex) {
-            ex.printStackTrace();
-            return null;
-        }
-    }
-}
