--- BukkitStaffChat/src/main/java/dev/imlukas/bukkitstaffchat/database/impl/sql/statement/LongStatementObject.java
@@ -1,22 +1,0 @@
-package dev.imlukas.bukkitstaffchat.database.impl.sql.statement;
-
-import java.sql.PreparedStatement;
-import java.sql.SQLException;
-
-public class LongStatementObject implements StatementObject {
-
-    private final long longNumber;
-
-    public LongStatementObject(long longNumber) {
-        this.longNumber = longNumber;
-    }
-
-    public static LongStatementObject create(long longNumber) {
-        return new LongStatementObject(longNumber);
-    }
-
-    @Override
-    public void applyTo(PreparedStatement statement, int index) throws SQLException {
-        statement.setLong(index, longNumber);
-    }
-}
