--- BukkitStaffChat/src/main/java/dev/imlukas/bukkitstaffchat/database/impl/type/DatabaseType.java
@@ -1,7 +1,0 @@
-package dev.imlukas.bukkitstaffchat.database.impl.type;
-
-public enum DatabaseType {
-
-    FETCHING,
-    CACHING
-}
