--- BukkitStaffChat/src/main/java/dev/imlukas/bukkitstaffchat/database/registry/DatabaseProvider.java
@@ -1,28 +1,0 @@
-package dev.imlukas.bukkitstaffchat.database.registry;
-
-import dev.imlukas.bukkitstaffchat.database.Database;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class DatabaseProvider {
-
-    private final List<Database> databases = new ArrayList<>();
-
-    public DatabaseProvider(List<Database> databases) {
-        this.databases.addAll(databases);
-    }
-
-    public static DatabaseProvider of(Database... databases) {
-        return new DatabaseProvider(List.of(databases));
-    }
-
-    public Database getDatabase(String name) {
-        for (Database database : databases) {
-            if (database.getName().equalsIgnoreCase(name)) {
-                return database;
-            }
-        }
-        return null;
-    }
-}
