+++ BukkitStaffChat/src/main/java/me/xii69/bukkitstaffchat/database/impl/sql/functions/ResultSetConsumer.java
@@ -1,0 +1,9 @@
+package me.xii69.bukkitstaffchat.database.impl.sql.functions;
+
+import java.sql.ResultSet;
+
+@FunctionalInterface
+public interface ResultSetConsumer {
+
+    void accept(ResultSet resultSet) throws Exception;
+}
