+++ BukkitStaffChat/src/main/java/me/xii69/bukkitstaffchat/database/impl/sql/statement/ObjectStatementObject.java
@@ -1,0 +1,24 @@
+package me.xii69.bukkitstaffchat.database.impl.sql.statement;
+
+import java.sql.PreparedStatement;
+import java.sql.SQLException;
+import java.sql.Types;
+
+public class ObjectStatementObject implements StatementObject {
+
+    private final Object object;
+
+    public ObjectStatementObject(Object object) {
+        this.object = object;
+    }
+
+    @Override
+    public void applyTo(PreparedStatement statement, int index) throws SQLException {
+        if (object == null) {
+            statement.setNull(index, Types.NULL);
+            return;
+        }
+
+        statement.setObject(index, object);
+    }
+}
