+++ BukkitStaffChat/src/main/java/me/xii69/bukkitstaffchat/database/impl/sql/statement/StatementObject.java
@@ -1,0 +1,11 @@
+package me.xii69.bukkitstaffchat.database.impl.sql.statement;
+
+import java.sql.PreparedStatement;
+import java.sql.SQLException;
+
+public interface StatementObject {
+
+    void applyTo(PreparedStatement statement, int index) throws SQLException;
+
+
+}
