+++ BukkitStaffChat/src/main/java/me/xii69/bukkitstaffchat/database/impl/type/DatabaseType.java
@@ -1,0 +1,7 @@
+package me.xii69.bukkitstaffchat.database.impl.type;
+
+public enum DatabaseType {
+
+    FETCHING,
+    CACHING
+}
